secp256k1: ported BitCurve.Add changes
Functional changes in func (BitCurve *BitCurve) Add in curve.go:
-(0,0) + (a, b) = (a,b).
-(a,b) + (a,b) == 2* (a,b) and not (0,0) anymore

Other changes to split out C code in order to be able to build fuzzers
